requirements.txt
Flask==1.1.2
Flask-MySQL==1.5.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
PyJWT==1.7.1